# http2: optimization of AppendHuffmanString